+++ src/handlers/reports/compliance_handler.py
@@ -90,13 +90,7 @@ class ComplianceReportHandler(AbstractHandler):
 
     @validate_kwargs
     def get_by_job(self, event: JobComplianceReportGetModel, job_id: str):
-        customer_name = event.customer or SystemCustomer.get_name()
-        job = self._job_service.get_by_customer_name(
-            customer_name=customer_name,
-            job_id=job_id,
-            job_types=event.job_types,
-        )
-        job = next(job, None)
+        job = self._job_service.get_nullable(job_id)
         if not job:
             return build_response(
                 content="The request job not found",